Transform links
Create `BlogAPI` class has the role of a wrapper for the Wordpress class.
This one will override the `get_articles` and `get_article` method and
apply the transformations required.

As a result the views file no longer needs to do any more
transformations every time it fetches the wordpress articles.

`BlogAPI` also accepts a boolean paramtere in the constructer
`transform_links`, if link transformations are wanted.